+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 10.3.1
+-----------------
+- Fix MySQL error result as comma separated list in Export.php
+
 Changes in 10.3
 ---------------
 - Add "Cumulative Balance over school years" checkbox in StudentBalances.php